Add support for M38 - Compute SHA1 hash of target file
Syntax is "M38 full/file/path.g". The printer then reads the file and computes its SHA1 hash, returning
that value as a single hexadecimal string with no trailing newline.

This command was added as part of a move towards higher automation - in particular, updating a file automatically
based on whether the replacement file has a different SHA1 hash. It is expected that external software implement the logic,
so the firmware need only concern itself with being able to provide information.

It is not recommended to issue an M38 whilst printing - it's quite computationally intensive.

A future enhancement would be to add a cache of results (perhaps SHA1 the filename -> SHA1 of file?),
as the results of this M code can take 30 seconds to compute (on an 11MiB file).
Note that the implementation used is a reference, rather than a highly optimised one. This could improve the runtime, but I
expect the limitation is I/O.