Add New 'C' Parameter to M301 To Enable GCode Control Of Heater Inversion
Due to the removal of the external FET pin on the Duet v0.8, we found that our external MOSFET-controlled
beds were no longer readily controlled by the board. Whilst it is possible to simply wire the external gate
up to the bed connector with a pull down to ground, this leads to the heater control being inverted.
As this only affects our heated bed, and not the hot end, changing the value of the HEAT_ON constant
would cause us issues elsewhere.

The solution is to let heater inversion be controlled by GCode - a new flag "C" has been added to M301
(Set PID parameters), which is read as a C-boolean (ie: 0 is false, anything else is true). If true,
then the heater will be inverted. This value is defaulted to the value of the HEAT_ON constant for all
heaters during the initialisation of the Heat class. The backing implementation is a bit-vector implemented
over an array of unsigned 8-bit integers (as this takes up the least space for the 7 heaters currently defined
for the Duet, which remains the primary target of this firmware). Calling classes need not concern themselves
with that detail, Heat::SetHeaterInversion(heater, inverted) and Heat::IsHeaterInverted(heater) provide readable
methods to get this data.